# openapi.yaml
openapi: 3.0.3
info:
  title: DOGFOOD API
  version: 0.1.0
  description: 'Hackathon judging, voting, and management API. Note that admin/service routes typically wrap responses in { data: ... } while some public routes return unwrapped JSON. Error formats also vary between wrapped structured objects and raw string messages.'
servers:
  - url: 'http://localhost:3000'
    description: Local deployment
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    JudgeAssignment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hackathon_id:
          type: string
          format: uuid
        judge_user_id:
          type: string
          format: uuid
        submission_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
    JudgeInvitation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hackathon_id:
          type: string
          format: uuid
        submission_id:
          type: string
          format: uuid
        judge_user_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - PENDING
            - ACCEPTED
            - DECLINED
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Policy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
        version:
          type: integer
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Criterion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        min_score:
          type: integer
        max_score:
          type: integer
        weight:
          type: integer
        display_order:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PolicyWithCriteria:
      allOf:
        - $ref: '#/components/schemas/Policy'
        - type: object
          properties:
            criteria:
              type: array
              items:
                $ref: '#/components/schemas/Criterion'
    Evaluation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hackathon_id:
          type: string
          format: uuid
        judge_user_id:
          type: string
          format: uuid
        submission_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
            - DRAFT
            - SUBMITTED
        total_score:
          type: number
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    EvaluationScore:
      type: object
      properties:
        criterion_id:
          type: string
          format: uuid
        raw_score:
          type: integer
    EvaluationDetail:
      type: object
      properties:
        evaluation:
          $ref: '#/components/schemas/Evaluation'
        scores:
          type: array
          items:
            $ref: '#/components/schemas/EvaluationScore'
...
paths:
...
  /api/judging/assignments:
    get:
      summary: Get judge assignments
      description: List active assignments for the authenticated user. Implicitly handles Judge isolation.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of assignments
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/JudgeAssignment'
      operationId: getapijudgingassignments
  /api/judging/invitations:
    get:
      summary: Get judge invitations
      description: List pending invitations for the authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Array of invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/JudgeInvitation'
      operationId: getapijudginginvitations
  '/api/judging/invitations/{id}/accept':
    post:
      summary: Accept judge invitation
      description: Transitions invitation to ACCEPTED and creates an assignment.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Accepted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
      operationId: postapijudginginvitationsidaccept
  '/api/judging/invitations/{id}/decline':
    post:
      summary: Decline judge invitation
      description: Transitions invitation to DECLINED.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Declined successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
      operationId: postapijudginginvitationsiddecline
  '/api/judging/submissions/{submissionId}/evaluation':
    get:
      summary: Get submission evaluation
      description: Get the authenticated judge's evaluation state for a specific submission. Isolates judges so they cannot see peer scores.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: submissionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Evaluation object
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/EvaluationDetail'
      operationId: getapijudgingsubmissionssubmissionIdevaluation
    put:
      summary: Save or submit evaluation
      description: Saves criteria scores and optionally commits them to SUBMITTED state.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: submissionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                submit:
                  type: boolean
                  default: false
                scores:
                  type: array
                  items:
                    type: object
                    required:
                      - criterion_id
                      - raw_score
                    properties:
                      criterion_id:
                        type: string
                        format: uuid
                      raw_score:
                        type: integer
      responses:
        '200':
          description: Evaluation updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/EvaluationDetail'
      operationId: putapijudgingsubmissionssubmissionIdevaluation
...
  /api/policies:
    get:
      summary: List policies
      description: Retrieve list of judging policies. Requires Organizer access.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Policy'
      operationId: getapipolicies
    post:
      summary: Create policy
      description: Create a new judging policy. Requires Organizer access.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                  maxLength: 10000
                  default: ''
      responses:
        '201':
          description: Created policy
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Policy'
      operationId: postapipolicies
  '/api/policies/{id}':
    get:
      summary: Get policy
      description: Get details of a policy with its criteria. Requires Organizer access.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Policy details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/PolicyWithCriteria'
      operationId: getapipoliciesid
  '/api/policies/{id}/criteria':
    post:
      summary: Add criterion
      description: Add a scoring criterion to a DRAFT policy. Requires Organizer access.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - min_score
                - max_score
                - weight
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                  maxLength: 10000
                  default: ''
                min_score:
                  type: integer
                max_score:
                  type: integer
                weight:
                  type: integer
                  minimum: 0
                  maximum: 100
                display_order:
                  type: integer
                  default: 0
      responses:
        '201':
          description: Created criterion
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Criterion'
      operationId: postapipoliciesidcriteria
  '/api/policies/{id}/criteria/{criterionId}':
    put:
      summary: Update criterion
      description: Update an existing criterion in a DRAFT policy. Requires Organizer access.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: criterionId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - min_score
                - max_score
                - weight
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                  maxLength: 10000
                  default: ''
                min_score:
                  type: integer
                max_score:
                  type: integer
                weight:
                  type: integer
                  minimum: 0
                  maximum: 100
                display_order:
                  type: integer
                  default: 0
      responses:
        '200':
          description: Updated criterion
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Criterion'
      operationId: putapipoliciesidcriteriacriterionId
    delete:
      summary: Delete criterion
      description: Delete a criterion from a DRAFT policy. Requires Organizer access.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: criterionId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Criterion deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      message:
                        type: string
      operationId: deleteapipoliciesidcriteriacriterionId
  '/api/policies/{id}/publish':
    post:
      summary: Publish policy
      description: Locks the policy for judging. Must have criteria weighting to exactly 100. Requires Organizer access.
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Published policy
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Policy'
      operationId: postapipoliciesidpublish
...